feat(ai): Move the status line and fade in streamed words
Connecting, Thinking and the closing duration were three separate
mounts, so each state arrived as a cut. They are now one line that
carries its own swap: the outgoing copy leaves upward through a blur
while the incoming one rises from below, and the shimmer stops once the
model does.

Streamed words fade in from a blur instead of appearing. svelte-
streamdown types its easing as a keyword union and its blur is not a
prop, so the timing is overridden in CSS against the inline animation
it writes per word.

Both surfaces share this, the AI chat and the support agent.
