--- src/lib/components/prompt-kit/loader/reasoning-shimmer-loader.svelte
@@ -1,29 +1,0 @@
-<script lang="ts">
-	import { cn } from '$lib/utils';
-
-	interface Props {
-		text: string;
-		class?: string;
-		size?: 'sm' | 'md' | 'lg';
-	}
-
-	let { text, class: className, size = 'md' }: Props = $props();
-
-	const textSizes = {
-		sm: 'text-xs',
-		md: 'text-sm',
-		lg: 'text-base'
-	};
-</script>
-
-<div
-	class={cn(
-		'bg-[linear-gradient(to_right,var(--muted-foreground)_40%,var(--foreground)_60%,var(--muted-foreground)_80%)]',
-		'bg-size-[200%_auto] bg-clip-text font-medium text-transparent',
-		'animate-[shimmer_4s_infinite_linear] motion-reduce:animate-none',
-		textSizes[size],
-		className
-	)}
->
-	{text}
-</div>
