+++ src/lib/components/prompt-kit/loader/text-shimmer-loader.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import { cn } from '$lib/utils';
+	import ShimmerText from '$lib/components/motion/shimmer-text.svelte';
 
 	interface Props {
 		text?: string;
@@ -16,14 +17,4 @@
 	};
 </script>
 
-<div
-	class={cn(
-		'bg-[linear-gradient(to_right,var(--muted-foreground)_40%,var(--foreground)_60%,var(--muted-foreground)_80%)]',
-		'bg-size-[200%_auto] bg-clip-text font-medium text-transparent',
-		'animate-[shimmer_4s_infinite_linear] motion-reduce:animate-none',
-		textSizes[size],
-		className
-	)}
->
-	{text}
-</div>
+<ShimmerText {text} class={cn('font-medium', textSizes[size], className)} />
